Added NodeMCU custom firmware files
Note: After flashing the firmware, the board becomes incompatible with the Arduino IDE and may result in a crash. This is indicated by a crash log on the serial monitor. NodeMCU flashed with a custom firmware file is compatible with ESPlorer IDE and not Arduino IDE. Hence, to get the board functioning normally on the Arduino IDE, flash the original firmware on to the board.